+++ .github/scripts/addCalcitePackageLabel.js
@@ -35,5 +35,18 @@ module.exports = async ({ github, context }) => {
       repo,
       labels: [package],
     });
+
+    await github.rest.actions.createWorkflowDispatch({
+      owner,
+      repo,
+      workflow_id: "issue-monday-sync.yml",
+      ref: "dev",
+      inputs: {
+        issue_number: issue_number.toString(),
+        event_type: "SyncActionChanges",
+        label_name: package,
+        label_action: "added"
+      },
+    });
   }
 };